backend/requirements.txt
fastapi==0.115.6
httpx==0.28.1
openai==1.59.7
pillow==11.0.0
pydantic-settings==2.7.0
python-multipart==0.0.19
uvicorn[standard]==0.32.1
